+++ packages/spacecat-shared-data-access/src/models/audit/audit.model.js
@@ -68,6 +68,7 @@ class Audit extends BaseModel {
     PRERENDER: 'prerender',
     PRODUCT_METATAGS: 'product-metatags',
     PRODUCT_METATAGS_AUTO_SUGGEST: 'product-metatags-auto-suggest',
+    PRODUCT_METATAGS_AUTO_FIX: 'product-metatags-auto-fix',
     SUMMARIZATION: 'summarization',
     PAGE_TYPE_DETECTION: 'page-type-detection',
     FAQS: 'faqs',