+++ packages/spacecat-shared-data-access/src/models/configuration/configuration.model.js
@@ -34,8 +34,10 @@ class Configuration extends BaseModel {
   static JOB_INTERVALS = {
     NEVER: 'never', // allows to enable imports without scheduling them.
     EVERY_HOUR: 'every-hour',
+    TWENTY_PAST: 'twenty-past',
     DAILY: 'daily',
     WEEKLY: 'weekly',
+    EARLY_MONDAY: 'early-monday',
     EVERY_SATURDAY: 'every-saturday',
     EVERY_SUNDAY: 'every-sunday',
     FORTNIGHTLY: 'fortnightly',