+++ packages/spacecat-shared-data-access/src/models/scrape-url/scrape-url.schema.js
@@ -12,7 +12,7 @@
 
 /* c8 ignore start */
 
-import { isValidUrl } from '@adobe/spacecat-shared-utils';
+import { isObject, isString, isValidUrl } from '@adobe/spacecat-shared-utils';
 
 import SchemaBuilder from '../base/schema.builder.js';
 import ScrapeUrl from './scrape-url.model.js';
@@ -46,9 +46,22 @@ const schema = new SchemaBuilder(ScrapeUrl, ScrapeUrlCollection)
     required: true,
     validate: (value) => isValidUrl(value),
   })
+  .addAttribute('processingType', {
+    type: 'string',
+    required: true,
+    validate: (value) => isString(value),
+  })
+  .addAttribute('options', {
+    type: 'any',
+    validate: (value) => !value || isObject(value),
+  })
   .addAttribute('isOriginal', {
     type: 'boolean',
     default: true,
-  });
+  })
+  .addIndex(
+    { composite: ['url'] },
+    { composite: ['isOriginal', 'processingType', 'createdAt'] },
+  );
 
 export default schema.build();