+++ packages/spacecat-shared-data-access/src/models/site/config.js
@@ -22,6 +22,7 @@ export const IMPORT_TYPES = {
   ORGANIC_KEYWORDS_QUESTIONS: 'organic-keywords-questions',
   ORGANIC_TRAFFIC: 'organic-traffic',
   TOP_PAGES: 'top-pages',
+  AHREF_PAID_PAGES: 'ahref-paid-pages',
   ALL_TRAFFIC: 'all-traffic',
   CWV_DAILY: 'cwv-daily',
   CWV_WEEKLY: 'cwv-weekly',
@@ -134,6 +135,13 @@ export const IMPORT_TYPE_SCHEMAS = {
     limit: Joi.number().integer().min(1).max(2000)
       .optional(),
   }),
+  [IMPORT_TYPES.AHREF_PAID_PAGES]: Joi.object({
+    type: Joi.string().valid(IMPORT_TYPES.AHREF_PAID_PAGES).required(),
+    ...IMPORT_BASE_KEYS,
+    geo: Joi.string().optional(),
+    limit: Joi.number().integer().min(1).max(2000)
+      .optional(),
+  }),
   [IMPORT_TYPES.CWV_DAILY]: Joi.object({
     type: Joi.string().valid(IMPORT_TYPES.CWV_DAILY).required(),
     ...IMPORT_BASE_KEYS,
@@ -209,6 +217,12 @@ export const DEFAULT_IMPORT_CONFIGS = {
     sources: ['ahrefs'],
     enabled: true,
     geo: 'global',
+  },
+  'ahref-paid-pages': {
+    type: 'ahref-paid-pages',
+    destinations: ['default'],
+    sources: ['ahrefs'],
+    enabled: true,
   },
   'cwv-daily': {
     type: 'cwv-daily',
@@ -304,6 +318,9 @@ export const configSchema = Joi.object({
       }),
     ).optional(),
     outputLocation: Joi.string().required(),
+  }).optional(),
+  tokowakaConfig: Joi.object({
+    apiKey: Joi.string().required(),
   }).optional(),
   contentAiConfig: Joi.object({
     index: Joi.string().optional(),
@@ -410,6 +427,7 @@ export const Config = (data = {}) => {
   };
   self.getLlmoCdnlogsFilter = () => state?.llmo?.cdnlogsFilter;
   self.getLlmoCdnBucketConfig = () => state?.llmo?.cdnBucketConfig;
+  self.getTokowakaConfig = () => state?.tokowakaConfig;
 
   self.updateSlackConfig = (channel, workspace, invitedUserCount) => {
     state.slack = {
@@ -657,6 +675,10 @@ export const Config = (data = {}) => {
     state.cdnLogsConfig = cdnLogsConfig;
   };
 
+  self.updateTokowakaConfig = (tokowakaConfig) => {
+    state.tokowakaConfig = tokowakaConfig;
+  };
+
   return Object.freeze(self);
 };
 
@@ -671,4 +693,5 @@ Config.toDynamoItem = (config) => ({
   brandConfig: config.getBrandConfig(),
   cdnLogsConfig: config.getCdnLogsConfig(),
   llmo: config.getLlmoConfig(),
+  tokowakaConfig: config.getTokowakaConfig(),
 });