+++ packages/spacecat-shared-data-access/src/models/site/index.d.ts
@@ -47,6 +47,7 @@ export type IMPORT_TYPES = {
   readonly ORGANIC_KEYWORDS: 'organic-keywords';
   readonly ORGANIC_TRAFFIC: 'organic-traffic';
   readonly TOP_PAGES: 'top-pages';
+  readonly AHREF_PAID_PAGES: 'ahref-paid-pages';
   readonly TOP_FORMS: 'top-forms';
 };
 
@@ -59,7 +60,7 @@ export type IMPORT_SOURCES = {
   readonly GSC: 'google';
 };
 
-export type ImportType = 'organic-keywords' | 'organic-traffic' | 'top-pages' | 'top-forms';
+export type ImportType = 'organic-keywords' | 'organic-traffic' | 'top-pages' | 'top-forms' | 'ahref-paid-pages' ;
 export type ImportDestination = 'default';
 export type ImportSource = 'ahrefs' | 'google';