+++ packages/spacecat-shared-example/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [@adobe/spacecat-shared-example-v1.2.31](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-example-v1.2.30...@adobe/spacecat-shared-example-v1.2.31) (2025-11-04)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/helix-universal to v5.3.0 ([#1091](https://github.com/adobe/spacecat-shared/issues/1091)) ([13cbc72](https://github.com/adobe/spacecat-shared/commit/13cbc721f67c066948337faa6c6a4ea5b0c5ec9a))
+
 # [@adobe/spacecat-shared-example-v1.2.30](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-example-v1.2.29...@adobe/spacecat-shared-example-v1.2.30) (2025-09-16)